refactor features for my comp
