+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/models/my-company.config.ts
@@ -1,4 +1,5 @@
 import { MyCompanyRowConfig } from './my-company-row.config';
+import { MY_COMPANY_FEATURE } from './my-company.model';
 
 export interface MyCompanyConfig {
   /**
@@ -126,10 +127,11 @@ export interface MyCompanyConfig {
   createConfig?: MyCompanyConfig;
 
   /**
-   * Indicates which features should be tested in the current suite
+   * Indicates which features should be tested in the current suite.
+   * @example features: [MY_COMPANY_FEATURE.USER_PASSWORD]
    */
-  featureToggles?: string[];
-  
+  features?: MY_COMPANY_FEATURE[];
+
   /**
    * SUBCATEGORY: Config for updating entities from submenu.
    */
