+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/models/my-company.model.ts
@@ -31,3 +31,7 @@ export interface TestListOptions {
     collapseAll?: boolean;
   };
 }
+
+export enum MY_COMPANY_FEATURE {
+  USER_PASSWORD = 'userPassword',
+}
