+++ projects/storefrontapp-e2e-cypress/cypress/helpers/b2b/my-company/my-company.utils.ts
@@ -22,7 +22,7 @@ export function testMyCompanyFeatureFromConfig(config: MyCompanyConfig) {
     testListFromConfig(config);
     testCreateUpdateFromConfig(config);
     testAssignmentFromConfig(config);
-    testFeaturesFromConfig(config.featureToggles);
+    testFeaturesFromConfig(config);
   });
 }
